api/requirements.txt
fastapi>=0.135.0
uvicorn>=0.42.0
pydantic>=2.11.0
pydantic-settings>=2.13.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
anthropic>=0.50.0
httpx>=0.27.0
python-docx>=1.1.0
